Move helpers to a helpers dir to speed up knife loading
Knife loads any file in the knife directory every time any command is run. These files were not actual plugins and didn't need to be loaded on every run. Moving them into a helper directory allows us to load them in the deps block at the plugin runtime, but prevents them from loading during any old knife command. On my local machine this resulted in knife-windows going from 0.370523 to load to 0.003768000053 seconds. 98x faster by moving the files.
