+++ .rubocop.yml
@@ -1,18 +1,18 @@
 Layout/DefEndAlignment:
   Exclude:
-    - 'lib/chef/knife/winrm_knife_base.rb'
+    - 'lib/chef/knife/helpers/winrm_knife_base.rb'
 
 Layout/EndAlignment:
   Exclude:
-    - 'lib/chef/knife/winrm_knife_base.rb'
+    - 'lib/chef/knife/helpers/winrm_knife_base.rb'
 
 Lint/AmbiguousRegexpLiteral:
   Exclude:
-    - 'lib/chef/knife/winrm_knife_base.rb'
+    - 'lib/chef/knife/helpers/winrm_knife_base.rb'
 
 Lint/AssignmentInCondition:
   Exclude:
-    - 'lib/chef/knife/winrm_knife_base.rb'
+    - 'lib/chef/knife/helpers/winrm_knife_base.rb'
 
 Lint/Loop:
   Exclude:
@@ -20,9 +20,9 @@ Lint/Loop:
 
 Lint/UselessAssignment:
   Exclude:
-    - 'lib/chef/knife/winrm_session.rb'
+    - 'lib/chef/knife/helpers/winrm_session.rb'
     - 'lib/chef/knife/wsman_test.rb'
 
 Naming/HeredocDelimiterCase:
   Exclude:
-    - 'lib/chef/knife/winrm_knife_base.rb'
+    - 'lib/chef/knife/helpers/winrm_knife_base.rb'
