+++ lib/chef/knife/helpers/knife_windows_base.rb
