+++ lib/chef/knife/helpers/winrm_base.rb
