+++ lib/chef/knife/helpers/winrm_knife_base.rb
