+++ lib/chef/knife/helpers/winrm_session.rb
