+++ lib/chef/knife/helpers/winrm_shared_options.rb
