+++ lib/chef/knife/helpers/wsman_endpoint.rb
